Check for oc client instead of openshift rpm's
Containerized environments won't have openshift rpm's. So instead
of checking for atomic-openshift-client, we can check if oc is
installed. Usually oc client is installed only on the master node,
even if the oc client is installed on rest of the nodes,
pbench-ansible will make sure oc tool is installed only on the
master.