pbench server: fix race with prefix file.
The prefix file has to be handled by the first script in the chain,
i.e. pbench-dispatch. Previously, it was handled by
pbench-unpack-tarballs, but that now leads to a race, since unpacking
and other processing happens simultaneously.

Move the handling of the prefix file from pbench-unpack-tarballs to
pbench-dispatch. There is still some handling in
pbench-unpack-tarballs because the results path has to take the prefix
into account.

This had to be done in *two* places, because of code duplication: I
missed the second instance before. We need to refactor the code.

After review:
- Fix typo.