pbench-server-activate: relax hostname check
Some servers are known in the config file by a name different than
what `hostname -f` returns on that server.  In such cases, the sanity
check in create-crontab prevents the crontab from being created.

Relax the check so that it gets not only the name of the host but an
alternate "real" name from the config file. We compare the result of
`hostname -f` against both: if one of them matches, we produce a
crontab, but if both fail, we don't.