pbench-move-unpacked: work around job pool hang
This is a workaround for a job pool problem: it seems that a
job termination message gets lost, so occasionally the script hangs with
a job pool process trying to read the next job from the job pool
queue, but there *is* no next job. This seems to happen *only* when
there is no work to be done.

Check if there is anything to be done: if not, exit immediately
without creating the job pool.