pbench-register-tool-set: flexible specification of tool collection intervals
It has become necessary to allow different tools to have different
collection intervals. In particular, pidstat tends to collect a lot of
data, so it is desirable to set its collection interval to something
larger than that of the rest of the tools. The --interval command line
switch applies to all tools in the default set and the config file
method was similarly limited.

pbench-register-tool-set is set to obey first the command line option
spec, then the default interval specified in the config file and then
a hardwired default interval (currently 3 seconds). Modify this slightly
so that a tool-specific spec in the config file overrides all of these.

E.g. the current setting in the config file is

[pbench/tools]
default-tool-set = sar, iostat, mpstat, pidstat, proc-vmstat, proc-interrupts, turbostat
interval = 3

By adding the following to the config file:

[tools/pidstat]
interval = 30

the default interval for pidstat (and only for pidstat) is changed
to 30 seconds.

These settings can be overridden for individual tools by calling
pbench-register-tool with --interval, but providing sensible defaults
in the general case is important for the ease of use of pbench.

They can also be changed through a user-specific config file that
overrides some settings in the default config file. This will be
documented in the pbench user documentation.