pbench-fio: fix pre-iteration-script and targets option handling.
Fixes #718
Fixes #723

Multiple targets are mishandled in the generation of the fio
jobfile. This commit fixes the generation of the jobfile.

The pre-iteration-script option was left out of the getopts string, so
it is not recognized as a legitimate option (although the rest of the
implementation is fine). This commit adds the option to getopts.

Add a unit test for the option handling. A unit test for the jobfile
generation is TBD.