pbench-metadata-log: do not muck around with hostnames
Fixes #725.

pbench-metadata-log used to believe that there is one canonical
name for the host and that is what `hostname -s` would return.
That is not viable as mentioned in issue #725.

It now uses the names that the user supplied when registering the
remotes to ssh into those remotes. It retrieves what `hostname -s`
returns and stores it as an extra hostname-s field. The indexer will
have to be made a bit smarter a) to realize that this is not a tool
name and b) to use its value when correlating with what the sos report
contains.