Modify the pbench dispatch script to run as a service
pbench-dispatch is running as a cron job, scheduled every minute. That
adds a minute of latency to the processing. It's a simple script that
examines the TODO subdirs under each controller for tarball symlinks;
when it finds one, it checks its MD5 and dispatches it for further
processing.

Instead of running it as a cron job, we are now running it as a service:
constantly checking, using inotify, for new tarballs in all the TODO
subdirs. When it sees a new tarball, it will do exactly what it does now:
check MD5 and dispatch it appropriately.