pbench-move-unpacked: fix long-standing bug
pbench-unpack-tarballs is supposed to leave the following structure in place
for pbench-move-unpacked to handle:

- an unpacked tarball in tmp
- a symlink to it in $incoming
- a symlink to *that* in $results

When pbench-move-unpacked runs, it finds all the tarballs
in move-unpacked state and for each tarball, it would

- copy the unpacked tarball from tmp to $incoming.copy
- rm the $incoming symlink
- mv $incoming.copy $incoming
- change the state of the tarball

If all goes well, this works. But pbench-move-unpacked occasionally
got a failure in one of those steps that caused it to not finish
processing a tarball. It would then lather, rinse and repeat, but
would not be able to get out of the rut, because a step would fail,
the state would *not* be changed, and the same tarball(s) would be
handled again and again.

One such failure is caused by an initial run that succeeded in
creating the $incoming directory followed by some (unknown) failure
that would not change the state. That left pbench-move-unpacked
vulnerable to the infinite regress described above, because on entry
it always assumed that $incoming is a symlink, and it would try to
remove it with `rm`. But if $incoming is a directory, this would fail
and it could never get out of that rut.

Modify the handling to check if $incoming is a directory, a symlink
or something else: if it is a directory, `rm -rf` it; if it is a symlink,
`rm` it; if it's something else, fail.

That cures the symptom, but not the underlying cause: why was $incoming
a directory in the first place? That remains TBD.

In addition, most of the counters (number of errors in particular) are
useless in this context: they were decared at top level, incremented
in the function and used at the end. But the function is not called
directly: instead we use job_pool.sh which sets up a bunch of
pre-forked processes that call the function by communicating over a
FIFO with the main process.  Since they don't have any shared context,
this does not work. We enable the job log and should use that to find
out the fate of the sub-processes, but that remains TBD.

Instead of using the error counter to figure out whether to send mail
or not, we now use the temp file that accumulates the mail to send: if
it's empty we skip sending any mail out.

There was also a typo, a missing $ when cat'ing the mail_content file
into mailx, that is now fixed.