pbench-move-unpacked: fix bug.
A previously-thought "impossible" situation happened: even though
pbench-unpack-tarball should create an $incoming symlink (in the
case when we are unpacking in a local fs, before moving it to the
dfs afterwards), we have seen situations where $incoming has been
a directory and also now a situation where $incoming does not exist
at all. The directory case was handled in PR #755, which also added
the catch-all case that was deemed impossible.

This commit qualifies that case: if $incoming exists in some form
other than symlink or directory, we deem it impossible and return
an error. If it does not exist at all, we wonder why but we go on
and finish the processing (which creates it as a directory).