pbench-fio: put defaults in config file
Put default values in the config file: do not hardwire them into
the script.

For now, just the version and histogram_interval_msec are treated.
The rest will follow if everything works according to plan.

Also the rest of the scripts will be treated the same way at some
point.

Fix the unit tests.

Fixes issue #741.