index-pbench: hostname impedance matching with tools
Tools may be registered using a "public" name but when the results
are collected, pbench uses "hostname -s" (possibly augmented by
a host label) as part of the pathname where the results are stored.
These are stored in the metadata log under the "tools/<host>" section
where the host part may be an FQDN, an IP address or anything else
that allows the tool registration to proceed.

Make index-pbench do the necessary impedance matching to go from
one convention to the other in order to find the tool data.