pbench-trafficgen: Fix unit tests.
Fix perl backwards-incompatibility  in BenchPostprocess.pm.
It worked fine on RHEL7 but failed on F27. Now it works in
both. That allows the pbench-fio unit tests to pass.

Fix pbench-moongen unit test: pbench-moongen now uses a default $config
setting, which used to be empty.

Unrelated and opportunistically added: A pbench-fio unit test was
failing because the exit status was not accounted for.

pbench-trafficgen: Fix tool postprocessing unit tests.

BenchPostproces.pm now produces shorter uids and avoids a stutter.