proc-interrupts-postprocess: deal with partly numeric IRQ names
The IRG "names" that appear in proc-interrupts output are either
completely numeric or completely alphabetic (e.g. "NMI") on x86_64,
but they can be alphanumeric ("IPI0") on aarch64.

The test for that in proc-interrupts-postprocess was assuming the
first two cases and failed when it encountered the third one. The
fix is to anchor the match (/^\d+$/) at both ends, to make sure
that only numerics match.