Undo setting of PATH in pbench-base.sh.
In PR #823, PATH was set in pbench-base.sh in order to allow the
scripts to find the pbench-report-status script. That broke some
unit tests.

This commit undoes that change and adds the PATH setting to the
crontab (through pbench-server-activate-create-crontab).

It also changes the test-8 input file to add a missing script-dir
setting in the config file.