Add --pbench-pre hook for UB; move wrapper
Like --pbench-post, we add a hook to be run before tools are
started. This gives us a chance do work knowing the `benchmark_run_dir`
and `benchmark_results_dir` locations, but before tools are started to
avoid disturbing tool data.

We also move the invocation of the user-benchmark-wrapper script until
after the --pbench-post script is invoked so that we can move work
generating result data outside of the tool data collection.