Rework pbench-move-unpacked to not use prefix
We have reworked this code for pbench-move-unpacked to avoid knowing about
the prefix or user links, by just replacing the incoming symlink with the
newly copied directory.

We also add similar error handling and tar ball gathering logic as we have
in pbench-unpack-tarballs.

The unit tests were updated so that pbench-move-unpacked is included in the
"run all scripts" test phase.  The curl test-bin has been modified to be
sure all invocations of curl are captured.