Fix extracted column metadata when indexing
The extracted column metadata only needs to be pulled once per
identifier, but needs to be applied to each data value for each
"`klass`".  It was only being applied to the last "`klass`" processed
instead of to all "`klasses`".  This should not actually change anything
since we don't index CSV files using the `unify` handler with more than
one "`klass`".