Fix timestamp handling; move errors to counters
The timestamp handling assumes that all timestamps are in millisecond
units.  Looking at some of the samples in the agent side of the tree,
there are examples where .csv files have timestamps as floating point
values which look like they are using units of seconds.

This set of changes modifies the .csv file timestamp handling to first
convert the timestamp to a float and then divide by 1,000 to convert it
to seconds without losing any precision.  However, if the original
timestamp was in units of seconds, then the conversion will be wrong.
We detect that case by comparing the timestamps to the start of the
pbench run itself, counting those as errors (since they should have been
emitted as millisecond values).

As we added a counter for error handling, we also changed all the
asserts to counted error conditions, and other exception handling as
well.  This should let more data get indexed without having an error
stop the entire process.