Add pre-computed source IDs to each indexed doc
Before we can address performance issues with bulk indexing, we need to
pre-compute the source ID field for each document we index so that value can
be used during error handling to look up the original JSON document for
re-transmission if necessary.

While doing that work, we also added better error handling when reading files,
and added a set of counters to record problems encountered while indexing
data.