Check tool timestamps against run end
We now check the timestamps of generated data records against both the
start and end of a run as a way of making sure we have valid timestamps
for tool data.

If tool data timestamps are not within the bounds of the overall
benchmark run, there is either a problem with the way timestamps are
being generated, or the expected units of milliseconds is actually
either microseconds or seconds.