Remove the pbench-server config from test state files
We have been relying on the pbench-server configuration in the state
tar balls for the proper unit test environment, but ignoring the
server configuration file in the tree.  This causes problems with a
test makes an assumption about the server enviroment that is not met.

This change ensures the pbench server configuration from the tree in
explicitly put into place, and removes the "opt" subtree from the state
files that had them using the script below.

----

for tf in *.tar.xz; do
    echo $tf
    mkdir wrk
    cd wrk
    tar xf ../$tf
    if [ -d opt ]; then
        find opt -ls
        rm -rf opt
        tar Jcf ../$tf .
    fi
    cd ..
    rm -rf wrk
done