Parallelize archive directory checks; compare file lists
In an attempt to speed up the verification process we have parallelized
the checks of the archive (primary) and the backup directories.

In addition, instead of comparing the output of the md5sum commands, we
compare the content of the combined md5sum files (which are kept in
sorted order by making sure the list of controllers is sorted before
processing, and leveraging "ls"'s sorted output.