Clean up test state entirely between tests
Tests were leaving artifacts around between tests, which can lead to
the behavior of one test impacting another.  This commit modifies the
method of test execution to place everying in $_testroot/pbench for
the state files, and other minor changes to test state setup so that
it is properly recreated cleanly for each test.