Enhance the unit test run environment
We make a simple run-unittests script that anybody can run from the
root directory, and use that for Travis CI.  For Travis CI we were
expecting the LANG=C environment variable, so we move that into the
individual unit test scripts, adding LC_ALL=C to ensure output is
sorted consistently.

The latest version of SCL uses python 3.6, so we now use rh-python36
for the SCL prefix.