Unit tests for pbench-copy-results and shims.
Make the config file used by the unit tests more generic

I tried to make the config file as similar as possible to the example
server/pbench/lib/config/pbench-server-default.cfg file.  The main
addition is the directory setting for the unit tests, so that the
/pbench, /pbench-local and /opt hierarchies can be specified relative
to that directory, instead of being absolute. However, there are minor
discrepancies in order to avoid breaking the existing unit tests.

The unittests script itself got two additions: a listing of the
contents of the /pbench-local hierarchy in the output and a cleanup
of the same when resetting the state.

The shim unt tests were fixed to conform to the conventions of the
new config file (state/config/pbench-server.cfg).