Index into Elasticsearch with streaming_bulk() API
Instead of making a list of actions, which gobbles up tons of memory
when you have thousands, perhaps millions, of documents to index, we
now use the streaming_bulk() Elasticsearch helpers API to work in
chunks of 500 documents (the default).

We close [1] over the "actions" generator argument so that we can
keep track of the original source document that was sent in the bulk
request, so that when handling errors we can queue it appropriately
when the document needs to be retried due to some indexing failure.

The other major change is to use a request timeout for the bulk
operations that is 100,000 minutes. Really, we should not timeout
at all, but this is long enough for now. By timing out, we end up
resubmitting the same bulk request again, while the original bulk
request is potentially still be processed by Elasticsearch. This
can lead to a build up of bulk requests artificially. Instead, we
just need to wait patiently for a bulk request to complete. If a
connection gets dropped, we'll retry immediately.

When handling failures, we use a capped, expontenial backoff wait
algorithm, which uses the system's random number generator to get
a uniform set of random backoff values, between 0 and the upper
bound as it grows per backoff window.  A "window" is defined
as the time period when items are found on the retry deque.

We have created a "mocked-out" streaming bulk API for unit tests
and debugging the indexing work that allows the same mechanism
that is used when actually indexing into Elasticsearch to be used
in those environments.

Further, we have enhanced the checking of the stream of documents
generated to ensure there are no duplicates being emitted and
that they have well-formed Elasticsearch arguments.

We also:

  * Remove the _tmpdir global variable
  * Add the notion of an error file so that documents that
    can't be indexed into Elasticsearch can be saved for review
    later
    * We have added the notion of an external file to accept bad
      Elasticsearch documents as an argument to es_index()
  * Raise the logging levels for the elasticsearch1 and urllib3
    modules to FATAL, since we are handling the retry logic here
  * The _DEBUG global variable is now connected to the dbg
    parameter of es_index()
  * We pull the environment variable references out of the main()
    function so that we reduce the environmental assumptions of
    the main module, looking ahead to enabling true unit tests in
    the python sense

[1] http://ynniv.com/blog/2007/08/closures-in-python.html