Fix indexing work to improve error handling
We fix the default "read timeout" to be very large so that we don't
timeout prematurely and retry needlessly for template operations.

We have enhanced the error handling so that we don't always assume we
get a response object, and we dump out error messages that are
unexpected for now.

Further we also:
  * add support for counting the number of times we retry a bulk
    request
  * add the new es_put_template wrapper which properly handles the
    retries
  * pull the print() function calls out of es_index() so that the
    caller can decide what they want to emit, not have that inside
    the library
  * fix the missing "math" library module reference
  * fix how we mock out the timestamps to avoid global vs. local
    variable namespace conflicts
  * move the changes to the urllib3 and elasticsearch1 logging
    libraries to the code path were we actually use Elasticsearch