Fix bugs, clean up crontab output, enhance unit tests
Quote variables in conditional tests.

Rearrange the crontab output, so that env variables are list first
in the crontab, before all the rules. Simplify the creation of the
header in the crontab by assuming there is a single mailfrom spec
in the [pbench-server] section of the config file.

Enhance the unittests script for test-8 (the activate test) to include
the generated crontab in the gold output. Also add start/stop
delimiters. Get rid of obsolete path setting/restoring and do not
delete the PATH setting in the crontab from the output: it is now
universal.