Fix index names generated by pbench-report-status
Fixes #901.

Index names are generated from the timestamp by dropping everything
except the YYYY-MM-DD spec. In particular the capital T that separates
the date from the time in the timestamp is not a legal character for
an ES index name.

Add error checking and remove unit test specials from pbench-report-status.