Improve unpack tar balls error handling
This is not a complete fix for error handling by any means, but it does
mean we will detect more errors with symlinks in TO-UNPACK, and clean
up better on some of the error conditions.

We have improved the "find" commands used to fetch all the data about
the tar balls to be processed so that we don't have to use the "du"
command, and we still process all the symlinks, including the bad ones.

We have added a unit test which exercises all the functional code
paths of the pbench-unpack-tarballs code.  It does not yet come close
to getting coverage of all the minor error paths.

We have made changes to those error paths so that we can tell a bit
more about what result tar ball had those errors.

We now unpack the tar ball to a special unpack directory so that we
can be sure the top level of the tar ball only has the result name, and
ignore any other data, with the ignored data counted as a warning.

We have made the first argument optional.

If the incoming directory is the same as the unpack path we were given
in the first argument, we now set the link destination to MOVED-UNPACKED
instead of UNPACKED.

Lastly, if a result tar ball has already been unpacked into the incoming
directory (either the temporary or the real one), we now error out so
that we can track that error condition.