Initial framework for auditing the archives
This is by no means complete, and hopefully will serve as a start for
others to add to significantly.

The state directory contents are not audited at all.

The prefix files are not checked to see that their respective tar balls
exist.

The reception areas are not checked.

Auditing of the archive tree is now run after all unit tests.