Use server-activate for setting up unit tests
In order to support the use of server-activate, we added the notion of
a setup script, optionally provided per test, allowing for state to be
removed after it is activated.

We also removed the various infrastructure that supported testing the
activate environment, no longer needed since it is part of the unit
tests now.

We also remove the _PBENCH_SERVER_TEST where possible, and/or remove
the "dryrun" variable.