Ensure PATH has the pbench server bin dir in it
We need to ensure that the PATH variable has the proper bin directory in
its list for the pbench server code.  We don't need to do this in the
unit test environment, though.

We also use $PROG instead of $(basename $0), fixing up places where exit
is called before log_finish, adding log_exit to assist.