Clean up set-result-state script
The `pbench-server-set-result-state` script was an attempt to address an issue
with the clients knowing too much about how the server side works.  With the
v0.49 release, this script was released into the wild, and was finally removed
with the v0.51 release.

For compatibility we have to maintain it, so we just clean up the comments and
left-over code.

We also remove the inotify reference from the `pbench-server-prep-shim-001`
script, along with the comment about duplicate code. The code duplication is
necessary given the need to maintain compatibility going forward.