fio-postprocess: more fixup to process fio options
Options from fio-result.txt are converted into
benchmark parameters for result.json.  This fixes
the per-job (in a job) options and concatenates those
together into a single benchmark-parameter, like:

  rw=read,read,read,write

The unittest output files for this are so old...

     -HOW OLD ARE THEY!?!

They are so old, there have been major version updates
of fio benchmark since the output files were added!
So, this update includes fio output that is actually
relevant, like output that includes the fio job options
we need to process - the whole point of this feature.