Index uperf result data properly.
This commit indexes sample result.json files from uperf and fio runs
properly, in the sense that the resulting documents do not include
arrays and include everything necessary so that the UI can retrieve
them and graph them.

TBD:

- Indexing of aggregated result.json files (at the iteration and
  global levels) is not addressed by this PR which concerns itself
  only with the lowest level result.json files (those in the
  <iteration>/sample-NN/ directories).

- Exceptions raised during generation of actions are handled at the
  wrong level: see issue #XXX for the planned future resolution of
  this problem.

New unit tests 7.15 and 7.16 address throughput and latency results
from uperf with small enough data sets so that the result data can be
examined in their entirety. Tests 7.8, 7.10, 7.11 contain larger data
sets (uperf and fio): the output contains only a subset of what would
actually be indexed, but the format can be examined to ensure its
adequacy.