Rewrite backup tarballs to support S3
This commit will do following things:
1. pbench-backup-tarballs.py: Add a new python script to
backup result to s3

This script will be invoked by the pbench-backup-tarballs.
After invoking it will check the TO-BACKUP state directory
for new results and figure out which are the results that
need to be backed up to s3.

2. pbench-verify-backup-tarballs.py: Add a new python script
for results verification

This script will be invoked by the pbench-verify-backup-tarballs.
The task will be verifying the results in archive, backup and
s3 bucket.

3. Add new configuration section for s3

4. Add new module for handling s3 configuration information

5. Fix Unit tests and dependencies