Require pbench-top|backup-dir in pbench-server.cfg
Remove pbench-top-dir and pbench-backup-dir from the default
configuration file so that they have to be explicitly defined for a
given pbench-server.  We then default pbench-local-dir to be the same as
pbench-top-dir.

Move BDIR definition out of pbench-base.sh because only
pbench-backup-tarballs and pbench-verify-backup-tarballs use it.

Update the pbench-server.cfg-example file to show that pbench-backup-dir
needs to be a directory and not a remote host reference, and add an
explicit definition of pbench-backup-dir to the unit test config files.

Update pbench-backup-tarballs and pbench-verify-backup-tarballs to
refine the error handling around the backup directory configuration.

Update a comment in pbench-server-activate-create-crontab to use an
example hostname, and use the satellite-prefix value in the lock file
name to avoid repeating ourselves.