Allow --defaults=list to not be interpreted as iterations One can run, "pbench-gen-iterations <benchmark> --defaults=list" which will provide a list of "defaults", a combination of parameters associated with a defaults name, such as "basic".  For example:
 > pbench-gen-iterations fio --clients=perf84 --defaults=list
  defaults available:
  basic: --rw=read,randread --filename=/tmp/fio-tst --filesize=256M --runtime=30s --bs=16k
  ceph-osp: --rw=read,randr,write,randw --bs=1m --ioengine=sync --numjobs=1 --runtime=30s

This works fine if calling pbench-gen-iterations directly, but if you
were to run pbench-run-benchmark <benchmark> --defaults=list, this script
will call pbench-gen-iterations, and the output (the list of defaults)
will be interpreted as iterations.  So, these changes prepend the defaults
output with "#", and pbench-run-benchmark will not interpret this as
iterations.