This is the default fio jobfile for using pbench-run-benchmarks
The jobfile used for pbench-fio, found in ./pbench-agent/bench-scripts/
templates, is not compatible because it is not in correct format for fio.
This new file is not in ./templates because it is not a template, but
a real fio jobfile.

This jobfile is the most basic jobfile one can use for fio, and
does not assume any options by default.  The user can either
specify fio parameters when running "pbench-run-benchmark fio ...",
or they can just provide a path to their own jobfile with --jobfile=.
If a user is not familiar with the various options for fio, they can
use one of the defaults via "--defaults=[basic]".  Otherweise, it
is up to the user to know and provide the proper options to make
fio run successfully.