Drop mailing status reports and use syslog
We no longer use mail for status reports if an Elasticsearch
instance is not found.  Instead we emit the JSON payload as
one line, and use send it along using syslog (via the 'logger'
command).  If a payload is greater than 4,096 bytes, we'll
just write it to a local timestamped file using xz compression.