Fixes: #700 parse pidstat in chunks.
when scan large pidstat-stdout.txt file, we might face OOM kill.
in order to avoid that, we should work in chunks, and pidstat
should hold just some buffer and not the entire file.

this patch will hold a buffer of 10,000 lines gen_data and reset the pidstat.

currently this method probably will take longer because we parse lines while scanning.
but this way we'll make sure we'll not hit any OOM kills.