Use ConfigParser instead of SafeConfigParser
As of Python 3.2 SafeConfigParser has been moved to ConfigParser.
As of Python 3.7 it appears that Python is now emitting warnings that
the deprecation has happened, and SafeConfigParser alias will be
removed in a future release.  So rather than having the warnings mess
up unittest output causing failures, let's just use ConfigParser only.

Note that we also remove the Python2 secondary import on failure.