Simplify PATH environment variable handling
First, pbench-base.sh was fetching the "script-dir" setting as BINDIR
but not using it, but the callers of pbench-base.sh were setting the
"dir" variable already which should be the same value.  So now we are
ensuring the CONFIG setting for "script-dir" matches the "dir" variable
where pbench-base.sh was invoked from.

Further, pbench-base.sh was updating the PATH variable to include the
path in the "dir" variable, so there is no need to have PATH set in
the crontab.

Finally, because we no longer have PATH in the crontab we don't need
the unit tests to strip down PATH just for their execution to avoid
test-8 (diffs expected crontab output) failing when the path changes.