A few cleanups to pbench-postprocess-tools
Based on a review for #1018, refactor a bit to clean up.

First and foremost, the handling of a missing non-default tool group has been fixed so that we emit an error and do nothing for a bad tool group.  This avoids collecting data under a tool group that does not exist.

We use `remote_hostname` everywhere instead of `remote` in certain places.

We added some comments and removed a level of indentation in one place that was not needed.