Gracefully handle closed indices
When we receive a closed index error, the operation type of the response
appears to change from "create" to "index".  So we first detect such
changes to be sure we can find the "error" payload.

Once we know we have a 403 status with the "IndexClosedException" in the
error payload, we store in the error file the full response and original
action for reprocessing later.