Fix mpstat post processing to handle timestamps properly
The headers of each table in the output contains a timestamp that needs
to be considered as well as those found with each data point.  When
that header is the first timestamp of the day, but also the last timestamp
before they roll over, if we don't consider it we'll be off by 24 hours
for all the other timestamps.